LCA-3: refactor: Implement ChatList component and update directory structure
- Restructured imports for Header and Main components to their respective folders.
- Refactored Main component to include a new ChatList component for better modularity and maintainability.
- Removed inline chat structure from Main, replacing it with ChatList.
- Added new CSS styles specific to ChatList.
- Removed old Main.css as styles have been refactored.
