+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
-import Header from "./components/Header";
-import Main from "./components/Main";
+import Header from "./components/Header/Header";
+import Main from "./components/Main/Main";
 
 const App = () => {
   return (
