--- frontend/src/components/Main.css
@@ -1,94 +1,0 @@
-main {
-  position: relative;
-  /* max-height: calc(100vh - 3.5rem); */
-  overflow: hidden;
-  flex-grow: 1;
-  display: flex;
-  align-items: stretch;
-  transition: all 0.35s;
-  section:not(:last-child) {
-    border-right: 1px solid rgb(var(--border));
-  }
-  section {
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: calc(100vh - var(--header-height));
-
-  }
-
-  #sidebar {
-    max-width: 40%;
-  }
-  #main {
-    flex-grow: 1;
-  }
-  /* https://www.w3schools.com/howto/howto_css_tooltip.asp */
-  .tooltip {
-    position: relative;
-    display: inline-block;
-  }
-  .tooltip .tooltiptext {
-    visibility: hidden;
-    width: 120px;
-    background-color: rgb(var(--border));
-    color: rgb(var(--text));
-    text-align: center;
-    font-size: 0.8rem;
-    border-radius: 0.5rem;
-    padding: 0.5rem;
-    position: absolute;
-    z-index: 1;
-    bottom: 125%;
-    right: -0.5rem;
-    opacity: 0;
-    transition: opacity 0.3s;
-  }
-  .tooltip .tooltiptext::after {
-    content: "";
-    position: absolute;
-    top: 100%;
-    right: 0.6rem;
-    margin-left: -5px;
-    border-width: 5px;
-    border-style: solid;
-    border-color: rgb(var(--border)) transparent transparent transparent;
-  }
-  .tooltip:hover .tooltiptext {
-    visibility: visible;
-    opacity: 1;
-  }
-
-  button {
-    padding: 0.5rem 1rem;
-    border: none;
-    cursor: pointer;
-    border-radius: 0.5rem; /* Rounded corners */
-  }
-  button[type="submit"] {
-    color: rgb(var(--background));
-    background-color: rgb(var(--primary));
-  }
-  button[type="reset"] {
-    background-color: rgb(var(--border));
-    background-color: rgb(var(--background));
-    /* color: rgb(var(--border)); */
-    border: 1px solid rgb(var(--border));
-  }
-}
-
-@keyframes slide-in-bottom {
-  from {
-    transform: translateY(100%);
-  }
-}
-
-@keyframes message-scale-out {
-  from {
-    transform: scale(0);
-  }
-  to {
-    transform: scale(1);
-  }
-}
