+++ frontend/src/components/Main/Main.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import ChatList from "../ChatList/ChatList";
 import "./Main.css";
 
 interface SectionProps {
@@ -17,42 +18,7 @@ const Main = () => {
   return (
     <main>
       <Section id="sidebar">
-        <div className="container-body">
-          <div className="tab">
-            <input type="checkbox" id="cb1" checked />
-            <div className="section-title">
-              <label htmlFor="cb1" className="tab-label">
-                My chats (2)
-              </label>
-            </div>
-            <div className="tab-content">
-              <div className="section-content">
-                <div className="section-column">
-                  <div className="chat-item">
-                    <div className="chat-item-avatar avatar-4">TC</div>
-                    <div className="chat-item-content">
-                      <p className="chat-item-title">Tina Cornell</p>
-                      <p className="chat-item-text">The test message</p>
-                    </div>
-                    <div className="chat-item-info">
-                      <span className="chat-item-time">15s</span>
-                    </div>
-                  </div>
-                  <div className="chat-item">
-                    <div className="chat-item-avatar avatar-7">CA</div>
-                    <div className="chat-item-content">
-                      <p className="chat-item-title">Collete Aicart</p>
-                      <p className="chat-item-text">Long messages should end with ellipses.</p>
-                    </div>
-                    <div className="chat-item-info">
-                      <span className="chat-item-time">35m</span>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
+        <ChatList />
       </Section>
       <Section id="main">
         <div id="chat-title" className="container-title">
